requirements.txt
annotated-types==0.5.0
anyio==3.7.1
exceptiongroup==1.1.2
fastapi==0.100.1
idna==3.4
pydantic==2.1.1
pydantic_core==2.4.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
alembic==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
boto3==1.26.42
botocore==1.29.42
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.0
fastapi-pagination==0.11.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
packaging==22.0
passlib==1.7.4
pipreqs==0.4.11
protobuf==3.20.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarg==0.1.9
